Converted WIRS to the new interface
